restructured as component
